reduce mangling of involved object names
The short_name logic seems to be targeted at pods, replicasets, and
statefulsets where parts of the name can be auto generated. In other
cases the short_name mangling is undesirable as it throws away relevant
parts of the name when the name contains dashes.  For example a Service
called "my-lb" would appear in the culprit only as "my".

This is an improvement but still imperfect because 1) it is quite
possible to create many of these objects directly without any generated
suffix and 2) even when a suffix is appended it is not always 2 segments
as currently assumed. A more complete fix likely requires querying
kubernetes for details about the involved object and determining the
closest object that is not automatically created by kubernetes to
attribute the culprit to.