Optionally only listen to events of a single namespace (#16)
In some setups one doesn't necessarily have the permission to listen to
events of the whole cluster. With multiple teams using a single cluster
not all events should go to the same sentry project.

By optionally settings the EVENT_NAMESPACE environment variable, only
events for that namespace will be reported.